requirements.txt
aniso8601==3.0.2
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
coveralls==1.7.0
Django==2.2
django-filter==2.1.0
django-graphql-jwt==0.2.1
docopt==0.6.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.8
promise==2.2.1
PyJWT==1.7.1
pytz==2019.1
requests==2.21.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3>=1.24.1
